+++ ratatoskr/obexagent.h
@@ -1,14 +1,15 @@
 /*==========================================================
  * Program : obexagent.h                 Project : ratatoskr
  * Author  : Michael Zanetti, Ian L., Philippe Andersson
- * Date    : 2025-12-18
- * Version : 0.0.1
+ * Date    : 2026-03-06
+ * Version : 0.0.2
  * Notice  : (c) Original work by Michael Zanetti, Canonical
  *           Adapted by Ian L. and Philippe Andersson
  * License : GNU GPL v3 or later
  * Comment : OBEX agent header
  * Modification History:
  * - 2025-12-18 (0.0.1) : Adapted from ubtd-20.04.
+ * - 2026-03-06 (0.0.2) : Added reject() for graceful transfer cancellation.
  *========================================================*/
 
 #ifndef OBEXAGENT_H
@@ -31,6 +32,7 @@ public:
     explicit ObexAgent(QObject *parent = 0);
 
     void accept(const QString & path, const QString &fileName);
+    void reject(const QString &path);
 
 signals:
     void authorized(const QString &path);