+++ Changelog.md
@@ -1,9 +1,13 @@
 # Changelog for project 'ratatoskr'
-(pan, 05/03/2026)
+(pan, 06/03/2026)
 
 
 ## v0.1.1 (*upcoming*)
 
+- Added confirmation pop-up to main 'ratatoskr' app to allow accepting or 
+  rejecting incoming file transfer (Github 
+  [issue #6](https://github.com/petroniusniger/ratatoskr/issues/6)).
+  Pop-up dialog shows file name and size.
 - Added `Changelog.md` (this file).
 - Sanitize path for incoming files in main 'ratatoskr' app (Github 
   [issue #12](https://github.com/petroniusniger/ratatoskr/issues/12)).